added a username, time and date to comments
Because the DB is not set up just yet, I have added a username input form on the new comment section. Additionally, I have created an array of objects that contain the username, the date and time of the comment, and the comment itself. Need to still add it to the comments section. Need to fix styling and how to output onto page