implemented array to handle more comments and lectures and also a page to display all lectures
For this commit, a great deal was accomplished. The array where lectures were stored was changed in order to handle multiple lectures. A new page was added that would display all of the current lectures for that given student. This page includes the lecture title and description and a button to click to view the lecture (which routes you to the lecture video page Jack and Erick created). After uploading a new lecture (by a professor), it would reroute to the lectures page and be visible on the list and can be clicked to view.