Bare bones of Comments
Added a new component that would handle the comments. I have three components for the comments. The module as a whole is the comments box and within the comment box you will find another module that has the comment list posted and also a comment form that would take in user input. Cannot access the array just yet but will continue working on it